// pkgs/harnica-lib/src/internals/parser/types.ts
export type NoteQuality =
  | [type: "quality", argument: "M" | "m"]
  | [type: "quality", argument: "M" | "m", argument: "7"]
  | [type: "tune", argument: "b" | "#", key: "5"]
  | [
      type: "omit" | "dim" | "aug" | "sus" | "add" | "tension",
      argument: string,
    ];